+++ SocialApp/CreatePostVC.swift
@@ -1,0 +1,166 @@
+//
+//  CreatePostVC.swift
+//  SocialApp
+//
+//  Created by Chase McElroy on 3/31/17.
+//  Copyright © 2017 Chase McElroy. All rights reserved.
+//
+
+import UIKit
+import SwiftKeychainWrapper
+import Firebase
+
+class CreatePostVC: UIViewController, UITextFieldDelegate, UIImagePickerControllerDelegate, UINavigationControllerDelegate {
+
+    @IBOutlet weak var imageAdd: UIImageView!
+    @IBOutlet weak var captionField: UITextField!
+    
+    var imagePicker: UIImagePickerController!
+    var imageSelected = false
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        self.captionField.delegate = self
+        imagePicker = UIImagePickerController()
+        imagePicker.allowsEditing = true
+        imagePicker.delegate = self
+        
+        let tap: UITapGestureRecognizer = UITapGestureRecognizer(target: self, action: #selector(CreatePostVC.dismissKeyboard))
+        view.addGestureRecognizer(tap)
+    }
+    
+    func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [String : Any]) {
+        if let image = info[UIImagePickerControllerEditedImage] as? UIImage {
+            imageAdd.image = image
+            imageSelected = true
+        } else {
+            print("CHASE: A valid image wasn't selected")
+        }
+        imagePicker.dismiss(animated: true, completion: nil)
+        
+    }
+
+
+    
+    @IBAction func postBtnTapped(_ sender: Any) {
+        guard let caption = captionField.text, caption != "" else {
+            let alertController = UIAlertController(title: "", message: "Please write a caption before posting", preferredStyle: UIAlertControllerStyle.alert)
+            let okAction = UIAlertAction(title: "Ok", style: UIAlertActionStyle.default) {
+                (result : UIAlertAction) -> Void in
+            }
+            alertController.addAction(okAction)
+            self.present(alertController, animated: true, completion: nil)
+            return
+        }
+        guard let img = imageAdd.image, imageSelected == true else {
+            let alertController = UIAlertController(title: "", message: "Please select a photo before posting", preferredStyle: UIAlertControllerStyle.alert)
+            let okAction = UIAlertAction(title: "Ok", style: UIAlertActionStyle.default) {
+                (result : UIAlertAction) -> Void in
+            }
+            alertController.addAction(okAction)
+            self.present(alertController, animated: true, completion: nil)
+            return
+        }
+        
+        if let imgData = UIImageJPEGRepresentation(img, 0.2) {
+            
+            let imgUid = NSUUID().uuidString
+            let metadata = FIRStorageMetadata()
+            metadata.contentType = "image/jpeg"
+            
+            DataService.ds.REF_POST_IMAGES.child(imgUid).put(imgData, metadata: metadata) { (metadata, error) in
+                if error != nil {
+                    print("CHASE: unable to upload to firebase storage")
+                    let alertController = UIAlertController(title: "", message: "Unable to upload image", preferredStyle: UIAlertControllerStyle.alert)
+                    let okAction = UIAlertAction(title: "Ok", style: UIAlertActionStyle.default) {
+                        (result : UIAlertAction) -> Void in
+                    }
+                    alertController.addAction(okAction)
+                    self.present(alertController, animated: true, completion: nil)
+                    return
+                } else {
+                    print("CHASE: Successfully uploaded image to Firebase Storage")
+                    let downloadURL = metadata?.downloadURL()?.absoluteString
+                    if let url = downloadURL {
+                        self.postToFirebase(imgUrl: url)
+                    }
+                    
+                }
+            }
+        }
+        dismissKeyboard()
+    }
+    
+    func postToFirebase(imgUrl: String) {
+        let user = KeychainWrapper.stringForKey(KEY_UID)!
+        let post: Dictionary<String, Any> = [
+            CAPTION_DB_STRING: captionField.text! as AnyObject,
+            IMAGEURL_DB_STRING: imgUrl as AnyObject,
+            LIKES_DB_STRING: 0 as AnyObject,
+            USER_DB_STRING: user as AnyObject,
+            POSTED_DATE: FIRServerValue.timestamp() as AnyObject
+        ]
+        
+        let firebasePost = DataService.ds.REF_POSTS.childByAutoId()
+        firebasePost.setValue(post)
+        
+        captionField.text = ""
+        imageSelected = false
+        imageAdd.image = UIImage(named: "add-image")
+        
+        dismiss(animated: true, completion: nil)
+    
+    }
+    
+    
+    @IBAction func selectPhotoTapped(_ sender: Any) {
+        present(imagePicker, animated: true, completion: nil)
+    }
+    
+    @IBAction func cancelBtnTapped(_ sender: Any) {
+        dismiss(animated: true, completion: nil)
+    }
+    
+    // MARK: KEYBOARD FUNCTIONS
+    // Move View
+    func moveTextField(textField: UITextField, moveDistance: Int, up: Bool) {
+        let moveDuration = 0.3
+        let movement: CGFloat = CGFloat(up ? moveDistance : -moveDistance)
+        
+        UIView.beginAnimations("animateTextField", context: nil)
+        UIView.setAnimationBeginsFromCurrentState(true)
+        UIView.setAnimationDuration(moveDuration)
+        self.view.frame = self.view.frame.offsetBy(dx: 0, dy: movement)
+        UIView.commitAnimations()
+    }
+    
+    // Keyboard shows
+    func textFieldDidBeginEditing(_ textField: UITextField) {
+        moveTextField(textField: textField, moveDistance: -250, up: true)
+    }
+    
+    // Keyboard is hidden
+    func textFieldDidEndEditing(_ textField: UITextField) {
+        moveTextField(textField: textField, moveDistance: -250, up: false)
+    }
+    
+    //presses return key
+    func textFieldShouldReturn(_ textField: UITextField) -> Bool {
+        captionField.resignFirstResponder()
+        return true
+    }
+    
+    // Hide keyboard when user touches outside keyboard
+    override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
+        self.view.endEditing(true)
+    }
+    
+    //Calls this function when the tap is recognized.
+    func dismissKeyboard() {
+        //Causes the view (or one of its embedded text fields) to resign the first responder status.
+        view.endEditing(true)
+    }
+    
+
+}
