+++ SocialApp/SignInEmailVC.swift
@@ -10,37 +10,18 @@ import UIKit
 
 class SignInEmailVC: UIViewController, UITextFieldDelegate{
 
+
     @IBOutlet weak var emailField: FancyField!
+    @IBOutlet weak var pwdField: FancyField!
+
     
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        // Do any additional setup after loading the view.
+        self.emailField.delegate = self
+        self.pwdField.delegate = self
     }
     
     
-    
-    func checkEmailAddress(email: String) {
-        // Check if Username exist
-        
-        
-        DataService.ds.REF_USERS.child(EMAIL_DB_STRING).observeSingleEvent(of: .value, with: { (snapshot) in
-            if snapshot.hasChild(email) {
-                print("we're good")
-                
-                self.performSegue(withIdentifier: "accountExists", sender: nil)
-            } else {
-                print("username doesn't exist")
-                self.performSegue(withIdentifier: "accountDoesNotExists", sender: nil)
-            }
-        })
-    }
-
-    @IBAction func nextBtnTapped(_ sender: Any) {
-        if emailField.text != "" {
-            checkEmailAddress(email: emailField.text!)
-        }
-        
-    }
 
 }
