--- SocialApp/SignInPasswordVC.swift
@@ -1,20 +1,0 @@
-//
-//  SignInPasswordVC.swift
-//  SocialApp
-//
-//  Created by Chase McElroy on 3/30/17.
-//  Copyright © 2017 Chase McElroy. All rights reserved.
-//
-
-import UIKit
-
-class SignInPasswordVC: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-
-
-}
